made copy constructor assignment operators and other improvments of the code

+++ _Project_GLFW_/WindowComponent.cpp
@@ -62,7 +62,7 @@ void Window::SetWindow(int width, int height, const char* title)
 		}
 		if (Input::GetStates(true,true,GLFW_MOUSE_BUTTON_LEFT))
 		{
-			std::cout << "we just pressed some mouse key at "<<" "<<Vector2F<float>::cursorPos() << " ";
+			std::cout << "we just pressed some mouse key at "<<" "<<Vector2F<int>::cursorPos() << " ";
 		}
 		if(Input::GetStates(true,false,GLFW_MOUSE_BUTTON_LEFT))
 		{
